The patch solves this problem, I hope...
Christopher Kings-Lynne wrote:
> I'm still getting ltree failures on 64bit freebsd:
>
> sed 's,MODULE_PATHNAME,$libdir/ltree,g' ltree.sql.in >ltree.sql
> gcc -pipe -O -g -Wall -Wmissing-prototypes -Wmissing-declarations -fpic -DPI
> C -DLOWER_NODE -I. -I../../src/include   -c -o ltree_io.o ltree_io.c -MMD
> ltree_io.c: In function `ltree_in':
> ltree_io.c:57: warning: int format, different type arg (arg 3)
> ltree_io.c:63: warning: int format, different type arg (arg 4)
> ltree_io.c:68: warning: int format, different type arg (arg 3)

Teodor Sigaev

+++ contrib/ltree/crc32.c
@@ -105,6 +105,6 @@ crc32_sz(char *buf, int size)
 	len = 0;
 	nr = size;
 	for (len += nr, p = buf; nr--; ++p)
-		_CRC32_(crc, TOLOWER(*p));
+		_CRC32_(crc, TOLOWER((unsigned int)*p));
 	return ~crc;
 }
